--- a/index.js
+++ b/index.js
@@ -4,7 +4,8 @@ module.exports = {
   name: 'loader.js',
 
   init: function() {
-    this._super.apply(this, arguments);
+    this._super.init && this._super.init(this, arguments);
+
     this.treePaths['vendor'] = 'lib';
   },