--- a/lib/loader/loader.js
+++ b/lib/loader/loader.js
@@ -12,6 +12,30 @@ var loader, define, requireModule, require, requirejs;
     requirejs: requirejs
   };
 
+  requirejs = require = requireModule = function(name) {
+    stats.require++;
+    return findModule(name, '(require)').module.exports;
+  };
+
+  function resetStats() {
+    stats = {
+      define: 0,
+      require: 0,
+      reify: 0,
+      build: 0,
+      modules: 0,
+      exports: 0,
+      ensureBuild: 0,
+      resolve: 0,
+      resolveRelative: 0,
+      findModule: 0,
+    };
+    requirejs._stats = stats;
+  }
+
+  var stats;
+  resetStats();
+
   loader = {
     noConflict: function(aliases) {
       var oldName, newName;
@@ -53,6 +77,7 @@ var loader, define, requireModule, require, requirejs;
   var defaultDeps = ['require', 'exports', 'module'];
 
   function Module(name, deps, callback, alias) {
+    stats.modules ++;
     this.id        = uuid++;
     this.name      = name;
     this.deps      = !deps.length && callback.length ? defaultDeps : deps;
@@ -75,6 +100,7 @@ var loader, define, requireModule, require, requirejs;
   };
 
   Module.prototype.exports = function() {
+    stats.exports ++;
     if (this.finalized) {
       return this.module.exports;
     } else {
@@ -98,6 +124,7 @@ var loader, define, requireModule, require, requirejs;
   };
 
   Module.prototype.reify = function() {
+    stats.reify++;
     var deps = this.deps;
     var dep;
     var reified = this.reified;
@@ -130,7 +157,9 @@ var loader, define, requireModule, require, requirejs;
   };
 
   Module.prototype.build = function() {
+    stats.ensureBuild++;
     if (this.state === FAILED || this.state === LOADED) { return; }
+    stats.build++;
     this.state = FAILED;
     this.reify()
     this.exports();
@@ -138,6 +167,7 @@ var loader, define, requireModule, require, requirejs;
   };
 
   define = function(name, deps, callback) {
+    stats.define++;
     if (arguments.length < 2) {
       unsupportedModule(arguments.length);
     }
@@ -171,11 +201,8 @@ var loader, define, requireModule, require, requirejs;
     throw new Error('Could not find module `' + name + '` imported from `' + referrer + '`');
   }
 
-  requirejs = require = requireModule = function(name) {
-    return findModule(name, '(require)').module.exports;
-  };
-
   function findModule(name, referrer) {
+    stats.findModule++;
     var mod = registry[name] || registry[name + '/index'];
 
     while (mod && mod.isAlias) {
@@ -189,7 +216,9 @@ var loader, define, requireModule, require, requirejs;
   }
 
   function resolve(child, name) {
+    stats.resolve++;
     if (child.charAt(0) !== '.') { return child; }
+    stats.resolveRelative++;
 
     var parts = child.split('/');
     var nameParts = name.split('/');
@@ -222,6 +251,7 @@ var loader, define, requireModule, require, requirejs;
   };
 
   requirejs.clear = function() {
+    resetStats();
     requirejs.entries = requirejs._eak_seen = registry = {};
     seen = {};
   };