--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "loader.js",
-  "version": "4.0.5",
+  "version": "4.0.6",
   "description": "loader.js =========",
   "main": "lib/loader/loader.js",
   "directories": {