… benchmarks/scenarios/ember-just-modules.js
… benchmarks/scenarios/ember.js