- [ ] polish
- [ ] ...
- [x] add full ember test (or something that demonstrates the diversity of imports)

cc @krisselden 

---

note it is marginally faster but the idea (still needs testing) is by doing isolating the recursive path and the non-recursive and ensuring no user-land code is invoked during the recurisve that may mitigate the overhead introduced by invoking deopting code during deep recursion.

some benchmarks, largely demonstrating that no regressions have occured.

before

```
benchmark
   file: ./benchmarks/scenarios/ember.js
   total: 123ms
   per op: 12.3ms
```

after
```
benchmark
   file: ./benchmarks/scenarios/ember.js
   total: 107ms
   per op: 10.7ms
```