@stefanpenner 2016-04-18
slightly  faster in node 0.12 (9.2ms)