@stefanpenner 2016-06-03
> what's left?

wanted to test with ember once it can use this loader