@stefanpenner 2016-06-09
released as v4.0.8